timeNow should be an int.  Need to mirror this in MSLoggerBase

+++ dummy_src/uk/org/smithfamily/mslogger/ecuDef/Megasquirt.java
@@ -102,7 +102,7 @@ public abstract class Megasquirt
      * How long have we been running?
      * @return
      */
-    protected double timeNow()
+    protected int timeNow()
     {
         return 0;
     }
